+++ src/core/ProgressIndicators/ProgressRadial/ProgressRadial.test.tsx
@@ -96,3 +96,15 @@ it('renders determinate small  ProgressRadial', () => {
   const spinnerFill = container.querySelector('.iui-fill') as SVGCircleElement;
   expect(spinnerFill.style.strokeDashoffset).toBe('44');
 });
+
+it('should render all sizes of ProgressRadial', () => {
+  const sizes = ['small', 'x-small', 'large'] as const;
+
+  sizes.forEach((size) => {
+    const { container } = render(<ProgressRadial size={size} />);
+    const spinner = container.querySelector(
+      `.iui-progress-indicator-radial.iui-determinate.iui-${size}`,
+    ) as HTMLElement;
+    expect(spinner).toBeTruthy();
+  });
+});